DOMA-3303 refactored ticket id page
