+++ apps/condo/domains/ticket/utils/clientSchema/Ticket.ts
@@ -2,7 +2,7 @@
  * Generated by `createschema ticket.Ticket organization:Text; statusReopenedCounter:Integer; statusReason?:Text; status:Relationship:TicketStatus:PROTECT; number?:Integer; client?:Relationship:User:SET_NULL; clientName:Text; clientEmail:Text; clientPhone:Text; operator:Relationship:User:SET_NULL; assignee?:Relationship:User:SET_NULL; details:Text; meta?:Json;`
  */
 
-import { get } from 'lodash'
+import { get, isUndefined } from 'lodash'
 import { generateReactHooks } from '@condo/domains/common/utils/codegeneration/generate.hooks'
 
 import { Ticket as TicketGQL } from '@condo/domains/ticket/gql'
@@ -61,7 +61,7 @@ function formValuesProcessor (formValues: ITicketFormState): TicketMutationType 
             } else if (formValues[key]) {
                 result[key] = { connect: { id: formValues[key] } }
             }
-        } else if (formValues[key]) {
+        } else if (!isUndefined(formValues[key])) {
             if (key === 'deadline') {
                 result[key] = formValues[key].toISOString()
             } else {
